ai-server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
openai>=1.40.0
python-dotenv>=1.0
httpx>=0.27
